scylla_node: _update_jmx_pid: don't throw exception if scylla-jmx.pid not found
This method is called from ndoe.stop() and we've seen cases where
this file was not found.

Move the call to a more specific place where we want to kill scylla-jmx
And since this is not critical at this point do not throw exception
if the file not found (we should have the pid from when starting
scylla-jmx).

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>