scylla_node: watch_rest_for_alive: wait for others to be considered normal token owners
It is not enough for node to know about other nodes' tokens,
as they might not be reflected in the token_metadata map.

After checking tokens, check also the `/storage_service/host_id` api
that provides a list of nodes that are normal token owners and ready
to be used by queries.

Refs https://github.com/scylladb/scylladb/issues/15146

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>